added useFusePools hook
