+++ src/components/pages/Fuse/FusePoolsPage.tsx
@@ -1,6 +1,6 @@
+import React from "react";
 import { Avatar, AvatarGroup, Link, Spinner, Text } from "@chakra-ui/react";
 import { Center, Column, Row, useIsMobile } from "buttered-chakra";
-import React, { useMemo } from "react";
 import { useTranslation } from "react-i18next";
 import { useRari } from "context/RariContext";
 import { useIsSmallScreen } from "hooks/useIsSmallScreen";
@@ -15,47 +15,15 @@ import { ModalDivider } from "components/shared/Modal";
 import { Link as RouterLink } from "react-router-dom";
 import FuseStatsBar from "./FuseStatsBar";
 import FuseTabBar, { useFilter } from "./FuseTabBar";
-import { useQuery } from "react-query";
 import { useTokenData } from "hooks/useTokenData";
-import Fuse from "fuse.js";
+
 import {
-  filterOnlyObjectProperties,
   filterPoolName,
 } from "utils/fetchFusePoolData";
+
 import { letterScore, usePoolRSS } from "hooks/useRSS";
 import { SimpleTooltip } from "components/shared/SimpleTooltip";
-
-export interface FusePool {
-  name: string;
-  creator: string;
-  comptroller: string;
-  isPrivate: boolean;
-}
-
-interface MergedPool {
-  id: number;
-  pool: FusePool;
-  underlyingTokens: string[];
-  underlyingSymbols: string[];
-  suppliedUSD: number;
-  borrowedUSD: number;
-}
-
-const poolSort = (pools: MergedPool[]) => {
-  return pools.sort((a, b) => {
-    if (b.suppliedUSD > a.suppliedUSD) {
-      return 1;
-    }
-
-    if (b.suppliedUSD < a.suppliedUSD) {
-      return -1;
-    }
-
-    // They're equal, let's sort by pool number:
-
-    return b.id > a.id ? 1 : -1;
-  });
-};
+import { useFusePools } from "hooks/fuse/useFusePools";
 
 const FusePoolsPage = React.memo(() => {
   const { isAuthed } = useRari();
@@ -94,81 +62,9 @@ export default FusePoolsPage;
 
 const PoolList = () => {
   const filter = useFilter();
-
-  const isMyPools = filter === "my-pools";
-  const isCreatedPools = filter === "created-pools";
-
   const { t } = useTranslation();
 
-  const { fuse, rari, address } = useRari();
-
-  const { data: _pools } = useQuery(
-    address + " fusePoolList" + (isMyPools || isCreatedPools ? filter : ""),
-    async () => {
-      const [
-        {
-          0: ids,
-          1: fusePools,
-          2: totalSuppliedETH,
-          3: totalBorrowedETH,
-          4: underlyingTokens,
-          5: underlyingSymbols,
-        },
-        ethPrice,
-      ] = await Promise.all([
-        isMyPools
-          ? fuse.contracts.FusePoolLens.methods
-              .getPoolsBySupplierWithData(address)
-              .call({ gas: 1e18 })
-          : isCreatedPools
-          ? fuse.contracts.FusePoolLens.methods
-              .getPoolsByAccountWithData(address)
-              .call({ gas: 1e18 })
-          : fuse.contracts.FusePoolLens.methods
-              .getPublicPoolsWithData()
-              .call({ gas: 1e18 }),
-
-        rari.web3.utils.fromWei(await rari.getEthUsdPriceBN()),
-      ]);
-
-      const merged: MergedPool[] = [];
-      for (let id = 0; id < ids.length; id++) {
-        merged.push({
-          // I don't know why we have to do this but for some reason it just becomes an array after a refetch for some reason, so this forces it to be an object.
-          underlyingTokens: underlyingTokens[id],
-          underlyingSymbols: underlyingSymbols[id],
-          pool: filterOnlyObjectProperties(fusePools[id]),
-          id: ids[id],
-          suppliedUSD: (totalSuppliedETH[id] / 1e18) * parseFloat(ethPrice),
-          borrowedUSD: (totalBorrowedETH[id] / 1e18) * parseFloat(ethPrice),
-        });
-      }
-
-      return merged;
-    }
-  );
-
-  const filteredPools = useMemo(() => {
-    if (!_pools) {
-      return undefined;
-    }
-
-    if (!filter) {
-      return poolSort(_pools);
-    }
-
-    if (isMyPools || isCreatedPools) {
-      return poolSort(_pools);
-    }
-
-    const options = {
-      keys: ["pool.name", "id", "underlyingTokens", "underlyingSymbols"],
-      threshold: 0.3,
-    };
-
-    const filtered = new Fuse(_pools, options).search(filter);
-    return poolSort(filtered.map((item) => item.item));
-  }, [_pools, filter, isMyPools, isCreatedPools]);
+  const { filteredPools } = useFusePools(filter)
 
   const isMobile = useIsMobile();
